feat: clean up some of the abstraction
